requirements.txt
flask==2.2.5
werkzeug>=2.1.0
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0